scrum/requirements.txt
Django==1.11.18
django-filter==1.1.0
django-filters==0.2.1
djangorestframework==3.9.0
Markdown==3.0.1
psycopg2==2.7.6.1
psycopg2-binary==2.7.6.1
pytz==2018.9
